usb: gadget: uvc: configfs: add framebased format support
Summary:
add framebased format support
As we know, The framebased format can be used to support a lot of
formats other than YUV and MJPEG, for example: H264 or H265.
And Nowadays, the H264 and H265 is used more and more compared to the
MJPEG, so there is a need to support such usecase, although the new UVC
1.5 and the UVC1.1 extensions already support H264, but they are much
more expensive and most of the time, what we need is just a channel to
stream the video frames.

In this patch, we implement the framebased format and set it to H264 by
default, and the users could change the video format as they need through
the guid configfs attribute node.

Mainly porting from patch Ie1c62f5fb4fa
https://github.com/rockchip-linux/kernel/commit/de1b3d4cd77a19663a7a6144801dd7efc3c8eb08

Test Plan:
uvc configfs and uvc + hid + uac2 composite gadget ok.
(uvc ok with android mobile, hid & uac enumerate ok.)

Reviewers: fenfen.wang, zhenjian.he, zhaojun.li, huiliang.xu

Reviewed By: huiliang.xu

Subscribers: jianghe.xu

Differential Revision: https://cr.hobot.cc/D46755