media: ipu: bug [2726] ipu frame when 6-way yuv repeatedly kill work is null
Summary:
In the sif offline isp online ipu scenario, ddrin is the leader that will call the ipu's work.
When the ipu is closed first, clk will be closed. If the IPU's work is being executed at this time,
the access register may be blocked, so when the ipu exits, the frame must be set work to null

Test Plan: 6yuv repeatedly kill

Reviewers: shengqiang.gao, xingzhu.gong, xiaofeng.ling

Reviewed By: shengqiang.gao

Differential Revision: https://cr.hobot.cc/D73233