spi: fix panic when irq queue work on spidev releasing
Summary:
when spidev is releasing, irq could happens and queue work will
cause panic when the workqueue is destroyed.

Also the rx_assemble_work could be working when releasing, so add
a flag and lock to protect it.

This patch is back ported from rhode, 4 combined into 1.

Test Plan: No

Reviewers: xiaofeng.ling, yunqian.wang, siyuan.wang

Reviewed By: siyuan.wang

Differential Revision: https://cr.hobot.cc/D96246