Subject: bug [XJ3-*]set max_diff to 10000 and add injection data
Summary:
max_diff is a parameter used to process abnormal temperature data in pvt-ts module. In this patch,
it is changed from 20000 to 10000 to further improve the stability of system temperature control.
At the same time, the injection data is added into the module to test the code logic of exception handling.

Test Plan: pass

Reviewers: haipeng.yu

Reviewed By: haipeng.yu

Differential Revision: https://cr.hobot.cc/D98966