ddr_mon: mpu: task [XJ3-3498] modify the protection range1 and range3
Summary:
1, Enlager the protection range of MPU_RANGE1 by setting the __end_rodata
address to MPU_S_RANGE1 register.MPU can protect the init memory
section of kernel and the protection gap between MPU_RANGE0 and
MPU_RANGE1 had been removed.
2, Distinguish between x3 and j3 by checking the kernel memory layout.
The range3 protects the IO area and 0xC00000000 from the BPU on X3 platform.
But on J3 platform, this section can't be protected because the range3 had
been used to protect kernel/user memory section.

Test Plan:
access the the physical address in protection range by vio
or bpu and check whether the data corrupted or not.

Reviewers: haipeng.yu

Reviewed By: haipeng.yu

Differential Revision: https://cr.hobot.cc/D106104