net : ethernet : fix [XJ3-3727] underflow and jabber timeout error in transmit
Summary:
We read the phy status and accordingly program the MAC register(duplex mode,speed,port...) in xj3_adjust_link function.
In databook , we should do not change the configuration when the MAC is actively transmitting or receiveing.
so, we have to make sure the MAC(rx & tx) is disable when we call the xj3_adjust_link.

Signed-off-by: qinlong.wang <qinlong.wang@horizon.ai>

Test Plan: No plan

Reviewers: xiaofeng.ling, yunqian.wang, zutao.min, yunzhi.ma

Reviewed By: xiaofeng.ling, yunqian.wang

Subscribers: qinlong.wang

Differential Revision: https://cr.hobot.cc/D122336